Implemented when part of chording rule
The when part of the rule gets the list of all the neighbors via
accumulate, then from this list the number of flagged squares is
calculated and the list of covered squares is obtained. Next the
neighborBombsCount and the number of flagged squares are compared and,
if equal, a further check on the number of covered squares is performed.
Variables are then printed for debug purposes.